+++ tasmota/include/tasmota_types.h
@@ -665,7 +665,8 @@ typedef struct {
   // ----------------------------------------
 #endif // ESP32S3
 
-  uint16_t      light_pixels;              // 496
+  uint16_t      light_pixels : 15;         // 496
+  uint16_t      light_pixels_reverse : 1;  // 496
   uint8_t       light_color[LST_MAX];      // 498  LST_MAX = 5
   uint8_t       light_correction;          // 49D
   uint8_t       light_dimmer;              // 49E
@@ -786,8 +787,10 @@ typedef struct {
   uint8_t       web_color2[2][3];          // EA0  Needs to be on integer / 3 distance from web_color
   uint16_t      zcdimmerset[5];            // EA6
 
-  uint8_t       free_eb0[22];              // EB0  22 bytes
+  uint8_t       free_eb0[20];              // EB0  20 bytes
 
+  uint16_t      light_pixels_height_1 : 15;// EC4  Pixels height minus 1, default 0 (0 means 1 line)
+  uint16_t      light_pixels_alternate : 1;// EC4  Indicates alternate lines in Pixels Matrix
   uint8_t       shift595_device_count;     // EC6
   uint8_t       sta_config;                // EC7
   uint8_t       sta_active;                // EC8